requirements.txt
PyPDF2
PyMuPDF
sentence-transformers
chromadb
google-generativeai
python-dotenv
langchain-text-splitters
ragas
langchain-google-genai
streamlit
duckduckgo-search